+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.1.0] - [unreleased]
+## [1.1.0] - 2016-01-02
 
 ### Added
 - "High quality table" border configuration for LatexTableConverter. (#3)